Added 'eps' page size to set a dpi if the image does not specify one. This produces a better size than forcing to A4 when making an eps. Added a -m# option to set the margin. Use -m0 when making an eps. Changed calculations from float to double for more accuracy on large images.

+++ psimage.h
@@ -25,12 +25,12 @@ typedef struct {
   int      height;                /* rows				 */
   int      components;            /* number of color components		 */
   int      bits_per_component;    /* bits per color component		 */
-  float    dpi;                   /* image resolution in dots per inch   */
+  double   dpi;                   /* image resolution in dots per inch   */
   DATAMODE mode;                  /* output mode: 8bit, ascii, ascii85	 */
   long     startpos;              /* offset to jpeg data		 */
   BOOL     landscape;             /* rotate image to landscape mode?	 */
   BOOL     adobe;                 /* image includes Adobe comment marker */
 } imagedata;
 
-#define	DPI_IGNORE (float) (-1.0) /* dummy value for imagedata.dpi       */
-#define DPI_USE_FILE ((float) 0.0)/* dummy value for imagedata.dpi       */
+#define DPI_IGNORE   ((double) (-1)) /* dummy value for imagedata.dpi    */
+#define DPI_USE_FILE ((double) 0)    /* dummy value for imagedata.dpi    */
